refactor: move campaign check to AetherII to prevent crash from loading client classes
